Move CompareDistNonterminal to attic
Only used in a test that could just as well use rdFMCS.AtomCompare.CompareAnyHeavyAtom
